backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pyotp==2.9.0
qrcode[pil]==7.4.2

# Plaid
plaid-python==17.0.0

# OpenAI
openai==1.12.0
tiktoken==0.8.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-dateutil==2.8.2
pytz==2024.1
email-validator==2.1.0.post1

# Background Tasks
celery==5.3.6
redis==5.0.1

# Monitoring
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
